Rearrage tests.
