+++ internal/converter/data_converter_test.go
@@ -29,10 +29,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	commonpb "go.temporal.io/api/common/v1"
-	"go.temporal.io/sdk/internal"
 )
 
 func testDataConverterFunction(t *testing.T, dc DataConverter, f interface{}, args ...interface{}) string {
@@ -127,81 +124,4 @@ func TestToStrings(t *testing.T) {
 	}
 
 	require.Equal(t, want, got)
-}
-
-func TestDecodeArg(t *testing.T) {
-	t.Parallel()
-	dc := getDefaultDataConverter()
-
-	b, err := internal.encodeArg(dc, internal.testErrorDetails3)
-	require.NoError(t, err)
-	var r internal.testStruct
-	err = internal.decodeArg(dc, b, &r)
-	require.NoError(t, err)
-	require.Equal(t, internal.testErrorDetails3, r)
-
-	// test mismatch of multi arguments
-	b, err = internal.encodeArgs(dc, []interface{}{internal.testErrorDetails1, internal.testErrorDetails2})
-	require.NoError(t, err)
-	require.Error(t, internal.decodeArg(dc, b, &r))
-}
-
-func TestProtoJsonPayloadConverter(t *testing.T) {
-	pc := NewProtoJSONPayloadConverter()
-
-	wt := &commonpb.WorkflowType{Name: "qwe"}
-	payload, err := pc.ToPayload(wt)
-	require.NoError(t, err)
-	wt2 := &commonpb.WorkflowType{}
-	err = pc.FromPayload(payload, &wt2)
-	require.NoError(t, err)
-	assert.Equal(t, "qwe", wt2.Name)
-
-	var wt3 *commonpb.WorkflowType
-	err = pc.FromPayload(payload, &wt3)
-	require.NoError(t, err)
-	assert.Equal(t, "qwe", wt3.Name)
-
-	s := pc.ToString(payload)
-	assert.Equal(t, `{"name":"qwe"}`, s)
-}
-
-func TestProtoPayloadConverter(t *testing.T) {
-	pc := NewProtoPayloadConverter()
-
-	wt := &commonpb.WorkflowType{Name: "qwe"}
-	payload, err := pc.ToPayload(wt)
-	require.NoError(t, err)
-	wt2 := &commonpb.WorkflowType{}
-	err = pc.FromPayload(payload, &wt2)
-	require.NoError(t, err)
-	assert.Equal(t, "qwe", wt2.Name)
-
-	var wt3 *commonpb.WorkflowType
-	err = pc.FromPayload(payload, &wt3)
-	require.NoError(t, err)
-	assert.Equal(t, "qwe", wt3.Name)
-
-	s := pc.ToString(payload)
-	assert.Equal(t, "CgNxd2U", s)
-}
-
-func TestJsonPayloadConverter(t *testing.T) {
-	pc := NewJSONPayloadConverter()
-
-	wt := internal.WorkflowType{Name: "qwe"}
-	payload, err := pc.ToPayload(wt)
-	require.NoError(t, err)
-	wt2 := internal.WorkflowType{}
-	err = pc.FromPayload(payload, &wt2)
-	require.NoError(t, err)
-	assert.Equal(t, "qwe", wt2.Name)
-
-	var wt3 *internal.WorkflowType
-	err = pc.FromPayload(payload, &wt3)
-	require.NoError(t, err)
-	assert.Equal(t, "qwe", wt3.Name)
-
-	s := pc.ToString(payload)
-	assert.Equal(t, "{Name:qwe}", s)
 }
